Update README.md
Join these telegram groups to get a feel of how the bot works and see it in operation.  For now we are testing the sell strategy, but the buy strategy is working.

Join this https://t.me/joinchat/8qCzm-OdWfVhMzNk telegram group and get notifications that don't pass our criteria. The bot doesn't buy the coins.  

Join this https://t.me/joinchat/yZ_Yzza6dA42NzI0 telegram group and get notifications that PASS our criteria.  The bot proceeds to buy them.  

The group for legit coins will also give you notifications immediately the tokens are bought, you will also get notifications as price for the token varies. This variation forms the basis of whether the bot has to sell the coin at a profit. 

The algorithm for sell is complex but can be provided on requests via whatsapp :  +254795557216